requirements.txt.txt
streamlit==1.32.0
gspread==6.0.2
google-auth==2.27.0
pandas==2.1.4
openpyxl==3.1.2